scengen/requirements.txt
torch
Pillow
salesforce-lavis
